[AIRFLOW-5252] Move GCP Transfer to core (#5858)
This commit moves GCP Transfer from contrib to core.
For more information check AIP-21.
