+++ airflow/contrib/operators/gcs_to_gcs_transfer_operator.py
@@ -17,17 +17,17 @@
 # specific language governing permissions and limitations
 # under the License.
 """
-This module is deprecated. Please use `airflow.contrib.operators.gcp_transfer_operator`.
+This module is deprecated. Please use `airflow.gcp.operators.cloud_storage_transfer_service`.
 """
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.contrib.operators.gcp_transfer_operator import (  # noqa
-    GoogleCloudStorageToGoogleCloudStorageTransferOperator,
+from airflow.gcp.operators.cloud_storage_transfer_service import (  # noqa
+    GoogleCloudStorageToGoogleCloudStorageTransferOperator
 )
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.contrib.operators.gcp_transfer_operator`",
+    "This module is deprecated. Please use `airflow.gcp.operators.cloud_storage_transfer_service`.",
     DeprecationWarning,
 )
