+++ tests/gcp/operators/test_cloud_storage_transfer_service_system.py
@@ -18,7 +18,7 @@
 # under the License.
 import unittest
 
-from tests.contrib.operators.test_gcp_transfer_operator_system_helper import GCPTransferTestHelper
+from tests.gcp.operators.test_cloud_storage_transfer_service_system_helper import GCPTransferTestHelper
 from tests.contrib.utils.base_gcp_system_test_case import SKIP_TEST_WARNING, DagGcpSystemTestCase
 
 from tests.contrib.utils.gcp_authenticator import GCP_GCS_TRANSFER_KEY
